distractionModel/requirements.txt
# Create a new requirements.txt with version constraints
numpy==1.24.3
flask
flask-cors
opencv-python
dlib==19.24.8
tensorflow<2.14.0  # Versions before 2.14 typically work with NumPy 1.x
keras
Pillow
